Added post-destroy callback to object pool. Refactored surface animation code.

+++ src/content/assets/model_face.h
@@ -13,7 +13,7 @@ namespace Assets {
 
 class ModelFace {
 public:
-	size_t Material;
+	int Material;
 	FlagSet<FaceTypeFlag> Type;
 	GeometryMode Geo;
 	LightMode Light;
